requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
cacheout==0.7.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==3.0.4
django-scheduler==0.9.3
enum-compat==0.0.2
eventlet==0.21.0
flasgger==0.9.4
Flask==0.12.2
Flask-Cors==3.0.8
flask-jsontools==0.1.7
flask-nameko==1.4.0
future==0.18.2
greenlet==0.4.13
httplib2==0.17.0
icalendar==4.0.5
idna==2.6
importlib-metadata==1.6.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==3.2.0
kombu==3.0.37
MarkupSafe==1.0
mistune==0.8.4
mock==2.0.0
modbus-tk==1.1.0
nameko==2.8.4
nose==1.3.7
numpy==1.18.2
paho-mqtt==1.5.0
path.py==11.0
pbr==3.1.1
ply==3.11
PyJWT==1.6.1
pymongo==3.6.1
Pyomo==5.6.9
pyrsistent==0.16.0
pyserial==3.4
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2019.3
PyUtilib==5.8.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
wrapt==1.10.11
zipp==3.1.0
